node: define the id as a long integer
The C standard sets the maximum value for an object of type int
to 32767 or greater. Usually all compilers define the int as a
4 bytes integer, this sets the maximum value to 2147483647. To
be on the safe side use long integers, that are guaranteed to
have a maximum value of 2147483647 or greater.

+++ src/cell.cpp
@@ -442,7 +442,7 @@ std::vector<int> Cell::extract_vertex_neighs(const std::vector<int> &vertices, c
 
 			// Number of vertices owned by the interface
 			int nCommonVertices = 0;
-			const int *interfaceConnect = interface.get_connect();
+			const long *interfaceConnect = interface.get_connect();
 			for (int k = 0; k < interface.get_vertex_count(); ++k) {
 				for (int n = 0; n < nVerticesToFound; ++n) {
 					if (interfaceConnect[k] == get_connect()[vertices[n]]) {
